OCSADV-200: extract zipper methods from ValidityCorrection for use elsewhere
